start implementing purple soldier state machine

+++ UnityZeldaLTTP/Assets/Scripts/IDamageable.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 
 public interface IDamageable {
-   void TakeHit (float damage, Vector3 hitPoint, Vector3 hitDirection);
+   void TakeHit (int damage, Vector3 hitPoint, Vector3 hitDirection);
 
-    void TakeDamage (float damage);
+    void TakeDamage (int damage);
 }
